Добавила валидацию полей количество комнат-гостей
